+++ src/bda/plone/orders/browser/bookings.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
-from bda.intellidatetime import convert
+from Products.CMFPlone.utils import safe_unicode
+from Products.Five import BrowserView
 from bda.intellidatetime import DateTimeConversionError
+from bda.intellidatetime import convert
 from bda.plone.orders import interfaces as ifaces
 from bda.plone.orders import message_factory as _
 from bda.plone.orders import permissions
@@ -17,7 +19,6 @@ from bda.plone.orders.transitions import transitions_of_main_state
 from bda.plone.orders.transitions import transitions_of_salaried_state
 from decimal import Decimal
 from odict import odict
-from Products.Five import BrowserView
 from repoze.catalog.query import Contains
 from repoze.catalog.query import Ge
 from repoze.catalog.query import InRange
@@ -25,7 +26,6 @@ from repoze.catalog.query import Le
 from yafowil.base import factory
 from zExceptions import InternalError
 from zope.i18n import translate
-
 import datetime
 import json
 import plone.api
@@ -41,7 +41,6 @@ class BookingsDropdown(BaseDropdown):
             vendor_uids = [vendor_uid]
         else:
             vendor_uids = get_vendor_uids_for()
-
         return BookingData(
             self.context,
             booking=self.record,
@@ -166,20 +165,20 @@ class BookingsTable(BrowserView):
         bookings_quantity = self.render_bookings_quantity(colname, record)
         bookings_total_sum = self.render_bookings_total_sum(colname, record)
         value = \
-            '<tr class="group_buyable">' \
-            '<td colspan="13">' + '<p>' + title + '</p>' +\
-            '<span>' +\
+            u'<tr class="group_buyable">' \
+            u'<td colspan="13">' + u'<p>' + safe_unicode(title) + u'</p>' +\
+            u'<span>' +\
             translate(
                 _("bookings_quantity", default=u"Bookings quantity"),
                 self.request
             ) \
-            + ': ' + str(bookings_quantity) + '</span>' \
-            '<span>' +\
+            + u': ' + safe_unicode(bookings_quantity) + u'</span>' \
+            u'<span>' +\
             translate(
                 _("bookings_total_sum", default=u"Bookings total sum"),
                 self.request
             ) \
-            + ': ' + str(bookings_total_sum) + '</td></tr>'
+            + u': ' + bookings_total_sum + u'</td></tr>'
 
         return value